refactor: Activity 생성을 생성할 Activity에게 책임을 부여
