+++ app/src/main/java/woowacourse/movie/dto/ReservationDto.kt
@@ -4,8 +4,4 @@ import java.io.Serializable
 
 data class ReservationDto(val movie: MovieDto, val detail: ReservationDetailDto) :
     Dto,
-    Serializable {
-    companion object {
-        const val RESERVATION_KEY_VALUE = "reservation"
-    }
-}
+    Serializable
